Split general display code from game-specific code so it can be reused in other games

+++ Piece/Piece.h
@@ -13,10 +13,10 @@ struct Point
 
 struct Color
 {
-    int r;
-    int g;
-    int b;
-    int a;
+    uint8_t r;
+    uint8_t g;
+    uint8_t b;
+    uint8_t a;
 
     bool operator==(const Color& other) const;
     bool operator!=(const Color& other) const;
